cleanup and fixups

+++ apps/utils/cppUtils.cpp
@@ -22,11 +22,12 @@ void operator delete[](void *p)
     free(p);
 }
 
-void operator delete(void* p, unsigned int index)
+void operator delete(void* p, size_t index)
 {
     void* ptr = (void*)((uint32_t)p);
     free(ptr);
 }
+
 
 template <typename T>
 class UniquePtr {
